rewrite report generation
This commit rewrites most of the report generation code, allowing users
to filter by large monsters and attack types. This also means more types
of attacks are included in the stats, such as barrel bomb, wyvernfire,
installation, and marionette.

A consequence of this change is that large monsters can now show up in
the table if they deal damage to another large monster. These items are
ignored by default but an option has been added to display them in the
table if desired.