WIP add save/load functionality using JSON api
This commit takes advantage of the new JSON api in REFramework to
provide save and load functionality. Default settings and presets are
now loaded from separate data files, and user settings can be saved and
loaded from their own file.

This commit also adds some folders to the repository to make it clear
where everything needs to go when installing the script.