support hotkey modifiers shift, ctrl, and alt
This commit allows the user to hold a modifier key when setting their
hotkeys. This will ensure the hotkey does not activate when the modifier
is not held.

The inverse is also true: hotkeys will not activate if a modifier that
was not registered with the hotkey is currently held.

closes #4