improve initialization of game resources
This commit moves some initialization into the update loop so that we
can initialize resources as they become available. This helps to avoid
issues where resources are not yet available at the time the script is
being loaded, requiring scripts to be reset through REFramework before
the overlay can appear.