track dps based on time monsters spend in combat
This commit implements dps tracking based on the cumulative time that
monsters spend in combat. Only monsters selected by the filter will be
counted, and damage will be compared against the total time that ANY
filtered monster was in combat.

Note that a lot of monsters seem to stay "in combat" for a long time
after the fight actually ends.