first pass for rewrite of drawing code
This commit rewrites most of the code used to draw reports, allowing
users to specify the order of columns they want and their widths. This
change means that columns are now the default means of spacing out the
data, and it is no longer an option to just append each text element to
the previous one.