fix issue with training dummy not appearing
This commit fixes an issue where the dummy in the training hall would
not appear on the tracker after leaving and re-entering the area. This
was happening because the training hall stay time was not being reset
internally until a second or so had passed after entering the area. The
script however would update within this time frame thereby setting the
LAST_UPDATE_TIME to whatever time was in the counter when the player
left the area. This would delay the next update until that time had been
reached again.

I fixed this by simply adjusting the LAST_UPDATE_TIME to the current
quest time if it was too far in the future. This should cover any
similar cases that might pop up elsewhere.