clear data only at the proper times
This commit changes when data is cleared to simplify the code and make
this action more intentional. Data is now cleared only when entering
either a quest or the training hall. Some function hooks that were
previously helping with this task were removed. I also removed the
clearing of data in the basic update loop.

As a consequence of these changes, data will be retained in the table
after leaving one of these places. The table will only be cleared upon
reentry.

This fixes a bug where the training dummy would sometimes stay in the
dataset when entering a quest. A bug was also fixed where the dummy
would appear twice in the table when reentering the training hall.