reorganize settings window
This is probably going to annoy some people, but these changes should
make the settings window more logical to navigate. More common settings
should now be in a more prominent position towards the top of the main
window, and more complex or less popular settings were moved to the
bottom of the window.

The buttons to open the hotkey and filter windows were also moved into
the settings window to keep everything in one place.