add debug window
This commit adds a new window for debugging purposes. The only option
for now is to allow displaying the total amount of damage which the
selected bosses are missing, but is not displayed somewhere in the
table. In other words, this option displays the total amount of damage
that should have been credited to someone but wasn't for some reason.