improve handling of monster damage
This commit improve several aspects of monster damage.

I discovered that some large monsters have an attacker ID of zero. This
causes their attacks to be tracked as if belonging to player 1 (both
within and without marionette.) This has been fixed by assigning a new
attacker ID to these monsters.

I also found a function on the AfterCalcInfo_DamageSide object which
allows us to tell whether an attack was made from a marionette rider. I
took advantage of this here to simply assign all marionette attacks to a
fake 'wyvern riding' row. This can be enhanced in the future to allow
crediting these attacks to the rider or to the monster being ridden.

Finally, a small amount of code was added to the monster update hook to
record the monster's attacker ID to the boss cache. This allows us to
display the names of large monsters who appear in the table.