add marionette attacker type
This commit adds a new attacker type called 'marionette'. This is not a
real type in the game; attacks are assigned to this type when they are
indicated by the game as being marionette attacks.

This will allow us to choose at the time of report generation how we
want to distribute these types of attacks. Currently, marionette attacks
will now always be attributed to the monster that attacked. In the
future this can be expanded to attribute these attacks to either the
rider or to a 'Wyvern Riding' pseudo item in the report.

This commit also enables the 'monster' attacker type by default, since
the merging with player one was fixed.