improve tracking of mounted player damage
This commit enhances the tracking of marionette damage and allows for
the option to credit players with marionette damage dealt while they
were riding a monster.

We now keep track of the current rider in the monster update method.
This means that when marionette damage is added to a boss, we can check
to see if the monster dealing this damage is currently being ridden. In
this case, the marionette damage will instead be merged into a special
damage counter for the riding player, and the report generation will
look up these special counters and merge them into the appropriate
report items depending on configured options.