Make reframework-d2d plugin optional
This commit adds a flag in the script to disable the use of d2d. This
flag can be manually set by the user, and will be automatically set if
d2d is not detected by the script. This will allow users to run the mod
in environments where d2d doesn't work, such as Proton.

Warning messages will be displayed in the log and the settings window to
notify the user that the plugin is missing.