Refactor mhrise-coavins-dps.lua into modules
This commit applies a massive refactor to the main script, separating
out its core components into multiple lua modules. This should make it a
lot easier to maintain and improve this project.

The following modules are added:

* State
* Core
* Enum
* Data
* Report
* Draw
* Hotkey
* Ui
* Hook
* Export
* User options

A user options module was added, to help separate user-defined
settings from the rest of the code.

The unit tests were updated, and some initial work was done to
separate them into modules:

* Data