Fix error due to reading keyboard too early
A runtime error is occurring in the hotkey code on boot, when attempting
to read the keyboard. I'm assuming this is because we are trying to read
the keyboard too early. This error is harmless but pollutes the
REFramework error log.

This commit delays the use of all managed resources by one frame. This
does appear to fix the error at first glance, but I'm not confident this
is a perfect solution.