Support new buddy attacker types
Sunbreak added two new attacker types for buddies, one of which is the
new Felyne Fireworks secret support move. This commit fixes #34 by
adding support to the filter mechanism for these new attacker types.

This commit also fixes the separate or merge of these new types of
damage with the hunter.