Add support for i18n
This commit introduces a decent rewrite and refactoring to allow support
for translation. Translation files can be added to the translations
directory and will be loaded at runtime. Users can switch between
translations at will.

This commit also adds a CJK-compatible font to the repository so that
CJK characters can be displayed in the settings window.

This commit also renames 'attacker types' to 'damage types'.