Add option to reveal how much damage isn't shown
This option will add a row to the table that shows how much damage is
missing from the table. Damage revealed in this way may or may not
indicate a problem. If you have some things disabled, such as followers
or small monsters, their damage will appear in this counter even if the
mod is correctly tracking it.

In order to use this feature to test for attacks that do not get
captured, you need to make sure you have all filters enabled so that all
possible damage will be shown in the table.