Add option to show total damage
This option adds a row at the end of the table which shows the total
damage included on the report. This will only reflect the full amount of
damage that the monster took if all filters are enabled (and the mod is
working perfectly) or "Reveal how much damage isn't shown" is enabled.