Refactor addDamageToBoss method signature
This commit removes the many parameters for this function and replaces
them with a table. Initialize this table using initializeDamageInfo()
in the data module and set the values you want before passing it in.

This helps avoid having to update all of the calls for this function
when we add more types of damage, in exchange for having to allocate one
of these things every time you want to call it. There is probably an
even better way to do this but for now this should be a small
improvement.