Fix poison/blast damage
This commit uses a different method to detect when poison and blast
damage has been triggered. Previously this was done in the enemy update
method using the activation count, but this only increments when the
local player activates the effect. Now, we are using the onActivateProc
method which is on the effect's param object, which seems to work even
when remote players activate the effect.

We also calculate the poison damage upfront using the damage, interval,
and timer information. This works around not seeing the poison ticks
when we aren't host.

Fixes #21