Fix marionette damage not tracking correctly
Rider information was not being captured correctly which meant
marionette damage could not be credited to the player who was riding the
wyvern.

There were also other issues relating to the addition of the 'large
monsters' filter preventing the correct display of marionette damage.

This commit also fixes issues with the 'include large monsters' filter.