Improve options to automatically change visibility
This commit makes several improvements to the options that automatically
change the overlay's visibility. All related options will now just
toggle on or off the display when certain events occur, instead of
overriding the display when in certain states. This should make the
behavior of these settings more intuitive and easier to use.

An option to control whether the overlay is shown on boot was added.